requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=2.1.3
pillow>=10.1.0
python-dotenv==1.0.0
torch>=2.2.0
torchvision>=0.17.0
requests>=2.31.0
python-jose>=3.3.0
joblib>=1.3.2
matplotlib>=3.8.0
seaborn>=0.13.0
